refactored the code, changed file structure to support file based expo routing and added new HomeScreen
