+++ BingoChecker/checkers.ts
