+++ app/_layout.tsx
@@ -1,5 +1,8 @@
 import { Stack } from "expo-router";
 
 export default function RootLayout() {
-  return <Stack />;
+  return <Stack>
+    <Stack.Screen name='index' options={{headerTitle:'Home'}}/>
+    <Stack.Screen  name='game' options={{headerTitle:'Bingo Game'}} />
+  </Stack>;
 }
